Added initialize function for UI interaction
